Start thurough testing of SSL's
Upon the first failure we know the SSL is not 'trusted' by a browser, so we can do a consecutive test to try and get the actual SSL in these cases.

Start testing the URL for proper DNS in the URL class.
New exception type for InvalidUrl

(cherry picked from commit 40011d42291c156cde5a5612dcd7c2ffb113cec7)