Distinguish trust vs validity
A trusted SSL is one that a browser trusts to load content for, this is an SSL that does not give any warning or error.
A valid SSL is an SSL that matches the domain it's being tested with.

SSL's can be a mixture of Valid and Trusted so it's best to distinguish between these now.

(cherry picked from commit 4813f4669e97b51dbeb446a0217800aa49fb3ce1)